Create and Use SSL for SillyTavern
Added ability for launcher to generate and implement SSL for SillyTavern.

Auto-creates SSL with OpenSSL bundled with GIT.
After creation updates ST launch links to use the cert.
Added abilities to delete and regen cert.
Added links to read me on why self-signed certs are safe in this case and how to add to trusted.

+++ Launcher.bat
@@ -333,11 +333,20 @@ REM ############################################################
 :home
 title STL [HOME]
 cls
+
+set "SSL_INFO_FILE=%~dp0\SillyTavern\certs\SillyTavernSSLInfo.txt"
+set "sslOptionSuffix="
+
+REM Check if the SSL info file exists and set the suffix
+if exist "%SSL_INFO_FILE%" (
+    set "sslOptionSuffix= (With SSL)"
+)
+
 echo %blue_fg_strong%/ Home%reset%
 echo -------------------------------------------------------------
 echo What would you like to do?
-echo 1. Start SillyTavern
-echo 2. Start SillyTavern + Remote Link
+echo 1. Start SillyTavern%sslOptionSuffix%
+echo 2. Start SillyTavern + Remote Link%sslOptionSuffix%
 echo 3. Start Extras
 echo 4. Start XTTS
 echo 5. Update Manager
@@ -391,7 +400,7 @@ if "%choice%"=="1" (
     ) else (
         call :create_custom_shortcut
     )
-)   else if "%choice%"=="0" (
+) else if "%choice%"=="0" (
     exit
 ) else (
     echo [%DATE% %TIME%] %log_invalidinput% >> %log_path%
@@ -410,8 +419,19 @@ if %errorlevel% neq 0 (
     pause
     goto :home
 )
-echo %blue_bg%[%time%]%reset% %blue_fg_strong%[INFO]%reset% SillyTavern launched in a new window.
-start cmd /k "title SillyTavern && cd /d %~dp0SillyTavern && call npm install --no-audit && node server.js && pause && popd"
+
+REM Check if SSL info file exists and set the command accordingly
+if exist "%SSL_INFO_FILE%" (
+    for /f "skip=0 tokens=*" %%i in ('type "%SSL_INFO_FILE%"') do (
+        goto :sslPathsFound
+    )
+    :sslPathsFound
+    echo %blue_bg%[%time%]%reset% %blue_fg_strong%[INFO]%reset% SillyTavern launched with SSL in a new window.
+    start cmd /k "title SillyTavern && cd /d %~dp0SillyTavern && call npm install --no-audit && node server.js --ssl && pause && popd"
+) else (
+    echo %blue_bg%[%time%]%reset% %blue_fg_strong%[INFO]%reset% SillyTavern launched in a new window.
+    start cmd /k "title SillyTavern && cd /d %~dp0SillyTavern && call npm install --no-audit && node server.js && pause && popd"
+)
 goto :home
 
 :start_st_rl
@@ -424,10 +444,24 @@ if %errorlevel% neq 0 (
     pause
     goto :home
 )
-echo %blue_bg%[%time%]%reset% %blue_fg_strong%[INFO]%reset% SillyTavern launched in a new window.
-start cmd /k "title SillyTavern && cd /d %~dp0SillyTavern && call npm install --no-audit && node server.js && pause && popd"
-start "" "%~dp0SillyTavern\Remote-Link.cmd"
+
+REM Check if SSL info file exists and set the command accordingly
+if exist "%SSL_INFO_FILE%" (
+    for /f "skip=0 tokens=*" %%i in ('type "%SSL_INFO_FILE%"') do (
+        goto :sslPathsFoundRL
+    )
+    :sslPathsFoundRL
+    echo %blue_bg%[%time%]%reset% %blue_fg_strong%[INFO]%reset% SillyTavern launched with SSL in a new window.
+    start cmd /k "title SillyTavern && cd /d %~dp0SillyTavern && call npm install --no-audit && node server.js --ssl && pause && popd"
+    start "" "%~dp0SillyTavern\Remote-Link.cmd"
+) else (
+    echo %blue_bg%[%time%]%reset% %blue_fg_strong%[INFO]%reset% SillyTavern launched in a new window.
+    start cmd /k "title SillyTavern && cd /d %~dp0SillyTavern && call npm install --no-audit && node server.js && pause && popd"
+    start "" "%~dp0SillyTavern\Remote-Link.cmd"
+)
 goto :home
+
+
 
 :start_extras
 REM Run conda activate from the Miniconda installation
@@ -4256,13 +4290,28 @@ REM ############################################################
 :editor_core_utilities
 title STL [EDITOR CORE UTILITIES]
 cls
+set "SSL_INFO_FILE=%~dp0SillyTavern\certs\SillyTavernSSLInfo.txt"
+set "sslOption=2. Create and Use Self-Signed SSL Certificate with SillyTavern to encrypt your connection &echo       %blue_fg_strong%Read More: https://sillytavernai.com/launcher-ssl (press 9 to open)%reset%"
+
+REM Check if the SSL info file exists and read the expiration date
+if exist "%SSL_INFO_FILE%" (
+    for /f "skip=2 tokens=*" %%i in ('type "%SSL_INFO_FILE%"') do (
+        set "expDate=%%i"
+        goto :infoFound
+    )
+    :infoFound
+        set "sslOption=2. Regenerate SillyTavern SSL - %expDate% &echo     %blue_fg_strong%SSL NOTE 1: You%reset% %red_fg_strong%WILL%reset% %blue_fg_strong%need to add the Self-Signed Cert as trusted in your browser on first launch. How to here: https://sillytavernai.com/launcher-ssl (press 9 to open)%reset% &echo     %blue_fg_strong%SSL NOTE 2: To remove the SSL press 8%reset%"
+
+)
+
 echo %blue_fg_strong%/ Home / Toolbox / Editor / Core Utilities%reset%
 echo -------------------------------------------------------------
 echo What would you like to do?
 echo 1. Edit SillyTavern config.yaml
-echo 2. Edit Extras
-echo 3. Edit XTTS
-echo 4. Edit Environment Variables
+echo %sslOption%
+echo 3. Edit Extras
+echo 4. Edit XTTS
+echo 5. Edit Environment Variables
 echo 0. Back
 
 set /p editor_core_util_choice=Choose Your Destiny: 
@@ -4271,13 +4320,21 @@ REM ######## EDITOR CORE UTILITIES - FRONTEND ##################
 if "%editor_core_util_choice%"=="1" (
     call :edit_st_config
 ) else if "%editor_core_util_choice%"=="2" (
-    call :edit_extras_modules
+    call :create_st_ssl
 ) else if "%editor_core_util_choice%"=="3" (
-    call :edit_xtts_modules
+    call :edit_extras_modules
 ) else if "%editor_core_util_choice%"=="4" (
+    call :edit_xtts_modules
+) else if "%editor_core_util_choice%"=="5" (
     call :edit_env_var
 ) else if "%editor_core_util_choice%"=="0" (
     goto :editor
+) else if "%editor_core_util_choice%"=="8" (
+    goto :delete_st_ssl
+) else if "%editor_core_util_choice%"=="9" (
+    echo Opening SillyTavernai.com SSL Info Page
+    start "" "https://sillytavernai.com/launcher-ssl"
+    goto :editor_core_utilities
 ) else (
     echo [%DATE% %TIME%] %log_invalidinput% >> %log_path%
     echo %red_bg%[%time%]%reset% %echo_invalidinput%
@@ -4289,6 +4346,34 @@ if "%editor_core_util_choice%"=="1" (
 start "" "%~dp0SillyTavern\config.yaml"
 goto :editor_core_utilities
 
+:create_st_ssl
+call "%~dp0bin\create_ssl.bat" no-pause
+:: Check the error level returned by the main batch file
+if %errorlevel% equ 0 (
+    echo  %green_fg_strong%The SSL was created successfully.%reset%
+) else (
+    echo  %red_fg_strong%The SSL creation encountered an error. Please see \bin\SSL-Certs\ssl_error_log.txt for more info.%reset%
+)
+pause
+goto :editor_core_utilities
+
+:delete_st_ssl
+REM Check if the SillyTavern\certs folder exists and delete it if it does
+set "CERTS_DIR=%~dp0SillyTavern\certs"
+
+if exist "%CERTS_DIR%" (
+    echo %blue_fg_strong%Deleting %CERTS_DIR% ...%reset%
+    rmdir /s /q "%CERTS_DIR%"
+    if errorlevel 0 (
+        echo  %green_fg_strong%The SillyTavern\certs folder has been successfully deleted.%reset%
+    ) else (
+        echo  %red_fg_strong%Failed to delete the SillyTavern\certs folder. Please check if the folder is in use and try again.%reset%
+    )
+) else (
+    echo  %red_fg_strong%The SillyTavern\certs folder does not exist.%reset%
+)
+pause
+goto :editor_core_utilities
 
 
 REM Function to print module options with color based on their status
@@ -4603,7 +4688,7 @@ REM ############################################################################
 
 
 :edit_env_var
-rundll32.exe sysdm.cpl,EditEnvironmentVariables
+start "" rundll32.exe sysdm.cpl,EditEnvironmentVariables
 goto :editor_core_utilities
